# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce Medicine API",
    "version": "2.0.0",
    "description": "A comprehensive B2B RESTful API for Bulk Medicine Sales. Features include Role-Based Access Control (RBAC), multi-tier audit logging, dynamic tax management, promotional engine, and integrated customer/doctor credit systems.",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from the Auth login endpoint"
      }
    },
...
  },
  "paths": {
...
    "/settings": {
      "get": {
        "summary": "Get all system settings (Admin only)",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all settings"
          }
        }
      },
      "post": {
        "summary": "Create a new system setting (Super Admin)",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Setting created"
          }
        }
      }
    },
    "/settings/notifications": {
      "get": {
        "summary": "Get notification gateway settings",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notification configuration"
          }
        }
      },
      "put": {
        "summary": "Update notification gateway settings",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Settings updated"
          }
        }
      }
    },
    "/settings/email": {
      "get": {
        "summary": "Get SMTP/Email relay settings",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Email configuration"
          }
        }
      }
    },
    "/settings/sms": {
      "get": {
        "summary": "Get SMS gateway settings",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "SMS configuration"
          }
        }
      }
    },
    "/settings/{key}": {
      "get": {
        "summary": "Get a specific setting by unique key",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setting value"
          }
        }
      },
      "put": {
        "summary": "Update a specific system setting",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setting updated"
          }
        }
      },
      "delete": {
        "summary": "Permanent deletion of a setting key",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setting deleted"
          }
        }
      }
    },
    "/settings/bulk": {
      "post": {
        "summary": "Bulk update multiple settings at once",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Settings updated"
          }
        }
      }
    },
...
  }
}